+++ src/histogram/atomic_hist_int.rs
@@ -514,13 +514,13 @@ where T: Ord + Sub<T, Output=T> + Add<T, Output=T> + One + NumCast + Copy
 impl<T> HistogramPartition for AtomicHistogramInt<T> 
 where T: Clone + std::fmt::Debug
 {
-    fn overlapping_partition(&self, n: usize, overlap: usize) -> Result<Vec<Self>, HistErrors>
+    fn overlapping_partition(&self, n: NonZeroUsize, overlap: usize) -> Result<Vec<Self>, HistErrors>
     {
-        let mut result = Vec::with_capacity(n);
+        let mut result = Vec::with_capacity(n.get());
         let size = self.bin_count() - 1;
-        let denominator = n + overlap;
+        let denominator = n.get() + overlap;
 
-        for c in 0..n {
+        for c in 0..n.get() {
             let left_index = c.checked_mul(size)
                 .ok_or(HistErrors::Overflow)?
                 / denominator;
@@ -699,7 +699,7 @@ mod tests{
 
         for overlap in 0..=5 {
             for _ in 0..100 {
-                let n = uni_n.sample(&mut rng);
+                let n: usize = uni_n.sample(&mut rng);
                 let (left, right) = loop {
                     let mut num_1 = uni.sample(&mut rng);
                     let mut num_2 = uni.sample(&mut rng);
@@ -714,12 +714,12 @@ mod tests{
                         break (num_1, num_2)
                     }
                 };
-    
+                let n_nonzero = NonZeroUsize::new(n).unwrap();
                 let hist_fast = HistI8Fast::new_inclusive(left, right).unwrap();
                 let hist_i = HistI8::new_inclusive(left, right, hist_fast.bin_count()).unwrap();
         
-                let overlapping_f = hist_fast.overlapping_partition(n, overlap);
-                let overlapping_i = hist_i.overlapping_partition(n, overlap);
+                let overlapping_f = hist_fast.overlapping_partition(n_nonzero, overlap);
+                let overlapping_i = hist_i.overlapping_partition(n_nonzero, overlap);
 
                 if overlapping_i.is_err() {
                     assert_eq!(overlapping_f.unwrap_err(), overlapping_i.unwrap_err());
@@ -794,7 +794,7 @@ mod tests{
         let uni_n = Uniform::new_inclusive(2, 6);
         for overlap in 0..=5 {
             for _ in 0..100 {
-                let n = uni_n.sample(&mut rng);
+                let n: usize = uni_n.sample(&mut rng);
                 let (left, right) = loop {
                     let mut num_1 = uni.sample(&mut rng);
                     let mut num_2 = uni.sample(&mut rng);
@@ -809,9 +809,10 @@ mod tests{
                         break (num_1, num_2)
                     }
                 };
+                let n_nonzero = NonZeroUsize::new(n).unwrap();
                 let hist_fast = HistI8Fast::new_inclusive(left, right).unwrap();
                 let hist_i = HistI8::new_inclusive(left, right, hist_fast.bin_count()).unwrap();
-                let overlapping_i = hist_i.overlapping_partition(n, overlap).unwrap();
+                let overlapping_i = hist_i.overlapping_partition(n_nonzero, overlap).unwrap();
 
                 assert_eq!(
                     overlapping_i.last().unwrap().borders().last(),
@@ -837,7 +838,7 @@ mod tests{
         for binsize in 2..=7 {
             for overlap in 0..=5 {
                 for _ in 0..100 {
-                    let n = uni_n.sample(&mut rng);
+                    let n: usize = uni_n.sample(&mut rng);
                     let (left, right) = loop {
                         let mut num_1 = uni.sample(&mut rng);
                         let mut num_2 = uni.sample(&mut rng);
@@ -856,9 +857,10 @@ mod tests{
                             break (num_1, num_2)
                         }
                     };
+                    let n_nonzero = NonZeroUsize::new(n).unwrap();
                     let hist_fast = HistI16Fast::new_inclusive(left, right).unwrap();
                     let hist_i = HistI16::new_inclusive(left, right, hist_fast.bin_count() / binsize).unwrap();
-                    let overlapping_i = hist_i.overlapping_partition(n, overlap).unwrap();
+                    let overlapping_i = hist_i.overlapping_partition(n_nonzero, overlap).unwrap();
                     
                     assert_eq!(
                         overlapping_i.last().unwrap().borders().last(),